Update timeseries diagnostics to allow for more advected species
As of v11-02c, the number of advected species in the standard and benchmark
simulations has exceeded 150. This causes issues in the timeseries
diagnostics (ND48, ND49, ND50, ND51, ND51b) because numbers 151 and above
are reserved for non-species quantities. To avoid issues in the future,
all non-species quantities have been moved to numbers 501 and above.

NOTE: This will all be obsolete when the netCDF diagnostics become standard.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>