Implement setting species concentration diagnostic
The species concentration diagnostic data is stored in
State_Diag%SpeciesConc. It is set just prior to updating History in main
by calling new subroutine Set_SpcConc_Diagnostic implemented in
unitconv_mod. It may be best to move this routine somewhere else in the
future, ideally state_diag_mod, but this will require some moving of
other code as well due to compile dependencies.

Set_SpcConc_Diagnostic accepts a diagnostic id string and a pointer to a
4D array. The diagnostic id is used to get the output units for the
diagnostic from the state_diag_mod metadata. State_Chm%Species data is
converted to the metadata units for the diagnostic, the diagnostic
array is set equal to State_Chm%Species, and the State_Chm%Species array
is converted back to its original units.

Note that this routine is general such that any species concentration
diagnostic array can be passed (not just State_Diag%SpeciesConc) and
any metadata may be retrieved (not just 'SpeciesConc'). This allows the
user flexibility to create multiple species concentration output
diagnostics, e.g. snapshots at different points within the timestep or
the same snapshot but with different units.

The original units of State_Chm%Species do not matter, in part by using
wrapper routine Convert_Spc_Units in unitconv_mod, also updated and
expanded in this commit. This allows usage in both GCC and GCHP without
the need for any special handling or awareness of species units at the
time of archiving.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>