Add error check to make sure netCDF output will always be written out
We now make sure that the "File Write" interval (which is specified in
"HISTORY.rc by the "frequency" tag for instantaneous collections and
by the "acc_interval" tag for time-averaged collections) is not shorter
than the length of the simulation as specified in input.geos.

Also, we no longer pass the NymdB and NhmsB fields to the HISTORY_INIT
function, because we can get those from the Input_Opt structure.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>