Structural updates to enable State_Diag output in GCHP diagnostics
The DiagList defined in diagnostics_mod.F90 now includes fields
isSpecies and Species. isSpecies is a logical indicating if the
item in HISTORY.rc has a species name attached. Species is the
species name string, if any. The species suffix is now separated
from the rest of the HISTORY.rc item name by two underscores
rather than one. This simplifies the string parsing, specifically
the extraction of the name used to retrieve metadata. It also makes
the base diagnostic name more clear.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>