Update gckpp_HetRates.F90 in halogens mechanism to include isoprene updates
In v11-02c, gckpp_HetRates.F90 was updated to include aqueous isoprene
uptake. This was not in the version of gckpp_HetRates.F90 used for the
halogens chemistry mechanism.

Also make minor cosmetic updates for consistency with code in
Standard/gckpp_HetRates.F90 and to avoid text running onto next line.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>