Remove tabs in 1st HISTORY.rc read and allow diagnostic names without '_'
To remove tabs in the 1st read of HISTORY.rc, which occurs in
diagnostics_mod.F90 as part of the initialization of the diagnostics
list, a copy of utility routine CleanText is included in diagnostics_mod.
This function should be moved to common utilities file in the future.

Also in this commit: correctly set registryID and include in the DiagList
information about HISTORY.rc items. Print it during Print_DiagList. The
registryID information is only used by GCHP.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>