Fixed unit conversion problem caused by typo in the unit name
In routine Convert_Spc_Units (in GeosUtil/unitconv_mod.F90), we fixed
a typo in the CASE statement.  The OutUnit name was "Kg/kg dry" but
should have been "kg/kg dry" (all lowercase).  This typo was not
caught with ifort 11.  But gfortran 6.3 threw several error messages
and the run later died with a segfault.  This has now been fixed.

As a result, we now have improved the error trapping in routine
READ_GC_RESTART (in GeosCore/restart_mod.F90) so that we exit with
an error message if RC returns failure.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>