Now insert new data values as the head values of linked lists
We have changed the algorithms in:

  MetaHistContainer_Insert
  MetaHistItem_Insert
  MetaRegItem_Insert

So that new nodes are added at the head of the linked list.  This will
add items to the list in reverse order of specification, which will
keep diagnostics of the same type together.  The prior algorithms
always kept the head node constant, which resulted in the first
diagnostic item being split from others of the same type.  This is
mostly a cosmetic fix.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>