Declare State_Diag%SpeciesConc as an 8-byte floating point real
We now have declared the State_Diag%SpeciesConc array as REAL(f8),
i.e. 8-byte real, so that we can potentially use this quantity for
checkpointing purposes in the future.

We have added to Headers/registry_mod.F90 and History/histitem_mod.F90
the capability to keep pointers to 2D/8-byte and 3D/8-byte arrays.

We also added a new routine (Register_DiagField_R8_4D) to the interface
in Headers/state_diag_mod.F90.  This was necessary in order to register
the new 8-byte State_Diag%SpeciesConc field.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>