Bug fix: Now update fields that are REAL(f8) in HISTORY_UPDATE
In the HISTORY_UPDATE routine (in History/history_mod.F90), we had
omitted code to update HISTORY ITEMS that archive data from REAL(f8)
arrays.  This caused e.g. the State_Diag%SpeciesConc diagnostic (which
is declared as REAL(f8)) not to have been archived.  Thus, the values
for SpeciesConc were all zeroes.

Once this bug was fixed, we obtain values for the e.g. SpeciesConc
diagnostic that are essentially identical (within numerical noise)
to the equivalent bpch diagnostics.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>