Remove superfluous special-handling for TagO3 in vdiff_mod.F90
In routine VDIFFDR (GeosCore/vdiff_mod.F90), there were blocks of code
that did special handling for tagged O3 simulations.  But with the
advent of the species database, all tagged species are included in
the range 1 ..State_Chm%nDryDep.  (This was not the case prior to the
species database)  Because the extra IF blocks were just duplicating
work, we have removed them.

Also, we fixed a typo in the definition of nAdvect in INIT_STATE_DIAG
(in Headers/state_diag_mod.F90).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>