Set Update for time-avg collections to heartbeat dt by default
We have made a change that now better mimics the behavior of
the MAPL History component GEOS-Chem "Classic".  For time-
averaged collections, the File Write interval is now set from
the .frequency tag of HISTORY.rc.  The Update interval is now
by default set from the Heartbeat timestep (which is the
recommended setting).  But you can also use the "acc_interval"
setting to override this (not recommended, but maybe needed
for debugging etc.).

For instantaneous collections, the Update and File Write
intervals are both set from the ".frequency" tag in HISTORY.rc.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>