Improve error trapping for undefined collection attributes
In History_ReadCollectionData (in History/history_mod.F90), we now
have added a better error trapping in the case any of the collection
attributes (e.g. ".frequency", "duration", ".filename", etc.) do not
correspond to the list of defined collection names.  The list of
collections is printed out with an informative error message, and
then execution is halted.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>