Use averaging period start not end time by default in HEMCO diag filename
The bpch and netcdf GEOS-Chem diagnostic files use the start time of the
averaging period in the filename. The HEMCO diagnostic file, on the other
hand, uses end time unless DiagnOutTimeStamp is included in
HEMCO_Config.rc. This  default behavior is changed to use start time in
the absence of DiagnOutTimeStamp for consistency across all GEOS-Chem
diagnostic files.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>