Now make the registry lookup case-insensitive
Added modifications to history_mod.F90 and registry_mod.F90 to make
the searches for items in State_Met%Registry, State_Chm%Registry,
and State_Diag%Registry case-insensitive.

Mostly what we do is to save the desired item name in a temporary
variable which is then converted to uppercase. Then we compare this
uppercase value against the list of registered fields in
State_Chem%Registry, State_Met%Registry, and/or State_Diag%Registry
(which are converted to uppercase during the initialization phase).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>