Pass errors encountered in input_mod.F to the top, then stop run
We have removed instances of ERROR_STOP (which kills a GC run wherever
it is called) from input_mod.F and time_mod.F.  We now use GC_Error
(in Headers/errcode_mod.F90) to print a message and then return
out of the subroutine.  We keep returning until we reach main.F,
where the run is terminated with a call to ERROR_STOP.

Minor fix in NcdfUtil/m_do_err_out.F90: Flush the stdout buffer
after writing, so that the error will appear when using gfortran.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>