Set maximum dynamic timestep to 60s if not 4x5, 2x2.5, 1/4 or 1/8th deg res
Prior to this commit MAX_DYN was not set unless the grid was defined as
4x5, 2.2x5, 0.5x0.625, 0.25x0.3125, or EXTERNAL_FORCING. This causes a
previously unidentified error in GCHP which is typically built with
EXTERNAL_FORCING undefined. The error is now observable because of
recent updates to input_mod that improves error passing in the stack.
Replacing the EXTERNAL_FORCING condition with setting a general default
value if none of the other grids are defined corrects the issue.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>