Remove PrecipFracCount, RainFracCount, WashFracCount
We have removed the modifications that added counter arrays in
State_Diag for the PrecipFracLs, RainFracLs, WashFracLs fields.

We were not able to match the bpch diagnostics exactly, so for
now we have abandoned this effort.  We may return to it after
we have gotten further ont he benchmark diagnostics.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>