Improved error checking when using acc_interval with time-avg collections
In routine History_ReadCollectionData (in History/history_mod.F90),
we now combine the UpdateYmd and UpdateHms fields into a single
variable if acc_interval is used.  (Ditto for FileWriteYmd and
FileWriteHms).  This makes it easier to test if the update interval
is larger than the file write interval.

NOTE: This only happens when manually specifying the update interval
for time-averaged collections with acc_interval.  Otherwise, the
update interval is set to the "heartbeat" timestep.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>