Initial updates for optical depth dust diagnostics
Includes optical depth of mineral dust (total) as well as optical depth
per wavelength in input.geos per dust bin. Currently dust bin is treated
as a wildcard (DUSTBIN) and data for each wavelength is stored in a
separate array within State_Diag. Wavelength is not inserted into the
diagnostic name and metadata lookup table and is instead referred to as
WL1, WL2, etc. Preliminary code to insert the wavelength is commented out
since it requires reading input.geos within diagnostics_mod as part of
the initialization of diag_list. This is necessary so that the metadata
id includes the wavelength string prior to its first call in GCHP, which
is prior to initialization of Input_Opt.

The diagnostics as is have not been validated.

Also in this commit: do not pass grid parameters to initialization
routines of State_Met, State_Chm, and State_Diag. Use USE statements
instead since the parameters are available from CMN_Size_Mod.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>