Use input.geos radiation menu wavelengths in optical depth dust diag names
The binned diagnostics for optical depth dust will have the configurable
wavelength in their main name and the bin as a suffix following a single
underscore. The radiation wavelengths are stored in a module-level
variable in diagnostics_mod so that they are available to GCHP prior to
initialization of Input_Opt. The metadata for State_Diag arrays are now
retrieved via an if-else block rather than a select-case block to allow
dynamic definitions of metadata ids.

Also in this commit: update length of string shown when printing
state_diag registered names to show full names of longer named diagnostics.
Keep print of diaglist in main on for development of diagnostics.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>