Also activate OHconcAfterChem netCDF diag in CH4 specialty simulation
For the sake of completeness, we have now activated the archival of
OH after CH4 chemistry to netCDF diagnostic "OHconcAfterChem".  This
matches the behavior of the prior code, which archived the OH to
slot #1 of the ND43 bpch diagnostic.

We now allocate & register the State_Diag%OHconcAfterChem array
if it is either a fullchem or CH4 simulation.  Also modified diag3.F
to write ND43 output for CH4 simulations.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>