Abstract retrieval of diagnostics tag name where wildcards are used
This update expands subroutine Get_TagInfo to return a tag string
given a wildcard string (tagId) and an index with the set [1:nTags].
Get_TagInfo may also be used to retrieve the total number of tags given
a wildcard string. This subroutine is used when registering state_diag
arrays that contain a wildcard and when adding history items to a
history collection in the presence of a wildcard.

All new wildcards must be implemented in Get_TagInfo within state_diag_mod.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>