Prevent NaN in CHEM_SO2 parallel loop
In routine CHEM_SO2 (in GeosCore/sulfate_mod.F), the variable SO2_sr
was assigned a value in the IF block where the sulfate production rate
enhancement is computed.  This is the only place where SO2_sr was
initialized.  It was not set to zero in the ELSE block.

But SO2_sr was used again outside of the IF block (in the equation
that defines rSIV).  A geosfp_4x5_tropchem unit test resulted in the
code dying at the point where rSIV was defined.  Therefore, we suspect
that SO2_sr being uninitialized in the ELSE block is causing the
error.

For now, set SO2_sr to zero in the ELSE block; this will make rSIV
zero.  This is sufficient to pass unit tests.  However, we should
double-check with the developer to make sure that this is the
intended behavior.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>