In gckpp_Hetrates.F90, reference KHETI_SLA only if it is a UCX simulation
In KPP/Standard/gckpp_Hetrates.F90, the State_Chm%KHETI_SLA array is
always referenced.  But because KHETI_SLA is not allocated unless
the UCX mechanism is used, this will cause a seg fault in the tropchem
simulation.

The fix is to bracket the State_Chm%KHETI_SLA with an IF statement
that tests if UCX is activated.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>