Bug fix: use local variable numTags to set output nTags in Get_TagInfo
This update fixes a bug used for netcdf diagnostics. nTags is an optional
output variable of Get_Tag Info but number of tags is needed within the
subroutine even if it is not passed as an argument. New local variable
numTags is now used instead of nTags so that number of tags can always
be set.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>