Remove extra ND43 arrays for SOA quantities (LBRO2H, etc)
Several SOA species were added to the ND43 chemical production
diagnostic in the past.  But a recent update to the SOA mechanism
removed the code that added these to the AD43 diagnostic array.
Therefore we are also removing several related arrays from the
various routines.

Also now attempt to bracket as many ADxx arrays with C-preprocessor
blocks #if defined( BPCH_DIAG ), which will make it easier to disable
bpch output if so desired.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>