Bug fix: adjust indices for ND43 bpch diag to avoid OOB error
We changed the size of the AD43 array from 7 to 4 to reflect that
fact that the SOA species (i.e. LBRO2H, etc.) are no longer updated
in the ND43 diagnostic.  But we also needed to adjust the index for
the slots of the AD43 array so that they were in the range 1..4.

For backwards compatibility with v11-01, we still plot OH with tracer
index =1, HO2 = 3, O1D = 4, O3P = 5.  We modified code in gamap_mod.F
and diag3.F accordingly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>