Avoid array-out-of-bounds error when setting TINDEX for obsolete ND09 diagnostic
ND09 is an obsolete diagnostic and setting the maximum number of tracers
to N_ADVECT+PD09 was causing array-out-of-bounds errors in the TOMAS40
simulation, which now has 499 tracers (the max number of tracers is 500).
To avoid any issues, we set the maximum number of tracers for ND09 to PD09.
This has no impact on model output because this diagnostic is not used.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>