Minor bug fixes for ND64 diagnostic (FAST-JX fluxes)
The GAMAP category character for ND64 was FJX-FLUX-$, but this is greater
than the 8 characters allowed.  Now changed this to FJX-FLX$ in both
gamap_mod.F and diag3.F.

Also, in gamap_mod.F, we now use a format of "I2.2" when converting
the NN number to a string.  This will prefix the zero if NN is less
than 10.  This was leaving a space in the tracer name in the
tracerinfo.dat file that was being created by gamap_mod.F.

NOTE: For some reason, the ND64 diagnostic returns all zeroes unless
UCX is turned on.  Need to figure out why.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>