Avoid floating underflow in strat_chem_mod.F90
The array PLVEC%LOSS is REAL*4. To avoid floating underflow when dividing
by species concentration, we copy State_Chm%Species to a REAL*4 scalar
called SpcConc and make sure that is greater than zero. This code may be
replaced in v11-02e when we use UCX prod/loss rates for all species, so
this may be a temporaty fix.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>