Reconfigured diag1.F for better efficiency (plus other updates)
We have reconfigured the code in GeosCore/diag1.F so that all of the
computations are done within a big OpenMP parallel DO loop.  This is
more efficient and encloses more work.

We increased the number of GEOS-Chem timers by one so that we could
add a timer to track the speedup of code modifications.

We have also commented out calls to Check_Var_Trop, Check_Spc, and
Check_Spc_Nested, which are done on every dynamic timestep.  These
routines, which contain DO looops, are probably no longer necessary.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>