Clean up tagged CH4 code in mixing_mod.F
It is inefficient to repeat the same loop three times (once for all advected
species, once for total CH4 species, and once for tagged CH4 species). To
fix this, we now do the loop over grid boxes once and add IF statements
where needed to treat the total CH4 and tagged CH4 species differently in
order to apply soil absorption.

Also block off code specific to the tagged CH4 simulation with IF statements
for ITS_A_CH4_SIM and LSPLIT.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>