Improve mapping for J-value diagnostics (still needs a bit of work)
We now declare the State_Diag%JVal and State_Diag%JNoon fields
with a dimension of nPhotol+2, which allows for the 2 extra slots
for O3_O1D and O3_O3P (for UCX) or O3 and POH (for non-UCX).
Modified various routines to save the indices from FAST-J for
these reactions into these two extra slots.

Also now return more errors back up to the main program.

NOTE: We still will need to zero the noontime J-value arrays in
flexchem_mod. Also, we might need to add JNoonDaily and JNoonMonthly
so that we can zero them appropriately.  Leave to another commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>