Add daily-average and monthly-average J-value netCDF diagnostics
GeosCore/fast_jx_mod.F: Cosmetic changes

GeosCore/flexchem_mod.F: Add arrays JvCountDay, JvSumDay, JvCountMon,
JvSumMon and remove JvCount and JvSum.  These are needed to compute
the daily and monthly average of noontime J-values.  These need to
be kept separately.

GeosCore/state_diag_mod.F: Added and registered JNoonDailyAvg and
JNoonMonthlyAvg fields of the State_Diag object.  Also fixed an indexing
problem in Get_TagInfo for the nPhotol+1 and nPhotol+2 slots of
the J-value arrays (Jval, JNoonDailyAvg, JNoonMonthlyAvg).

NOTE: For simulations with UCX, we match (to within numerical noise)
the bpch output.  For tropchem simulations, we match all species
except the JO3 (nPhotol+1) and JPOH (nPhotol+2).  This is because the
netCDF diagnostics for these slots are attached AFTER the call to
PHOTRATE_ADJ in DO_FLEXCHEM, but the bpch diagnostics are attached
before those rates are modified.  We might have to define special
netCDF diagnostics to exactly match this.  A small price to pay for
extra clarity...

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>