Fix bug impacting dust radius obtained in RDUST_ONLINE
In the UCX simulation we’re accessing the wrong file for when obtaining
aerosol optical properties for dust. In FAST-JX we have routine RD_AOD to
read all the .dat files containing the optical properties for different
aerosols. Those files are defined as:

       DATA SPECFIL /"so4.dat","soot.dat","org.dat","ssa.dat","ssc.dat",
#if defined( UCX )
      ! Treating strat sulfate with GADS data but modified to match
      ! the old Fast-J values size (r=0.09um, sg=0.6) - I think there's
      ! evidence that this is too smale and narrow e.g. Deshler et al. 2003
      ! NAT should really be associated with something like cirrus cloud
      ! but for now we are just treating the NAT like the sulfate... limited
      ! info but ref index is similar e.g. Scarchilli et al. (2005)
      !(DAR 05/2015)
     &              "h2so4.dat","h2so4.dat",
#endif
     &              "dust.dat"/

Then in the dust code, we obtain the dust properties (e.g. aerosol radius)
from the values read from the dust.dat file.

         ERADIUS(I,J,L,N) = RDAA(N,6) * 1.0e-4_fp

         TAREA(I,J,L,N)   = 3.e+0_fp / ERADIUS(I,J,L,N) *
     &                      DUST(I,J,L,N) / MSDENS(N)

The issue here is that aerosol radius is always accessed with index 6 in
the second dimension, which refers to the 6th file read in RD_AOD. For the
tropchem simulation, dust.dat is the 6th file read, but for UCX h2so4.dat
is the 6th file read. Therefore, in the UCX simulation the properties for
H2SO4 will be used for dust aerosol radius, which in turn impacts aerosol
area. The aerosol radius and area fields are used in other routines,
including gckpp_HetRates.F90. So the wrong values for dust are passed to
the heterogeneous chemistry routines and this results in different het
rates for the UCX and tropchem simulations. The differences are especially
noticeable in regions of high dust concentrations (e.g. the Middle East
where we saw large differences in O3 concentrations).

The quick and easy fix is to change the line in the dust code to make sure
we obtain the data from the correct file:

#if defined ( UCX )
         ERADIUS(I,J,L,N) = RDAA(N,8) * 1.0e-4_fp
#else
         ERADIUS(I,J,L,N) = RDAA(N,6) * 1.0e-4_fp
#endif

We could eventually change the order of files listed in FAST-JX so that the
h2so4.dat files are listed at the end, but we won't do that here in case
there other places in the code that use hardwiring as in dust_mod.F to
access aerosol properties.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>