Remove UCX CPP switch and replace with Input_Opt%LUCX everywhere -- initial attempt
All uses of "#if defined ( UCX )" have been removed from GEOS-Chem and the
UCX flag has been removed from Makefile_header.mk. We now use Input_Opt%LUCX
instead, which allows us to switch UCX on and off without having to recompile
GEOS-Chem.

!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
Some issues still need to be resolved. GEOS-Chem simulations with UCX are
crashing because of variables defined in CMN_SIZE_mod.F and CMN_FJX_mod.F
that depend on Input_Opt%LUCX are set before input.geos is read, so the
non-UCX value is always used. This will also be an issue with GCHP because
GCHP requires parameters from CMN_SIZE_mod.F, etc. to initialize the
diagnostics at the start of the simulation. One solution may be to add a
module to read certain lines from input.geos at the very start of the
simulation and this can be called where needed in GCHP and GCC. An example
of this is currently in routine Init_DiagList in Headers/diagnostic_mod.F.
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>