Add minor fix for ND22 diagnostic
Lei Zhu reported an issue in the ND22 diagnostic. He wrote:

   GEOS-Chem crashes when ND22 diag is turned off. This is because
   LTJV is used in flexchem_mod.F90 (ling 623) no matter whether it
   has been defined (in ndxx_setup.F line 548) or not.

The quick fix is to move the allocation of LTJV to outside the "#if
defined ( BPCH_DIAG )" block so that it is allocated for both bpch
and netCDF diagnostics. However, note that LTJV is only computed in
diag_2pm.F if ND22 is turned on. That means if ND22 is turned off or
if BPCH_DIAG=no then LTJV will always be zero. We should eventually
fix this for the netCDF diagnostics.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>