Add fields of state met for querying vertical regime (#1)
Added State_Met%InChemGrid, State_Met%InTropopause, and
State_Met%InPbl fields which will eventually replace the various
inquiry functions.

As of this commit, the only one of these that have been activated
is State_Met%InTropopause.  We have verified that this works.

NOTE: In future commits we will remove the calls to the various
inquiry functions and replace them by these fields.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>