Use simple SOA as default in PM2.5 and AOD calculations; Remove LISOPOA option
Following the advise of the Aerosol Working Group, the code for computing
PM2.5 and OCPISOA in AEROSOL_CONC has been updated so that the simple
SOA species is used as default over the complex SOA species when both
options are turned on (e.g. in the benchmark simulations). OCPISOA
is used to compute AOD and also impacts the heterogeneous chemistry.
In addition, the option to turn on/off the mechanistic isoprene SOA
option added be Eloise Marais (LISOPOA) has been removed. Those
isoprene SOA species are always included in the chemical mechanisms,
so they should always be included in PM2.5 and AOD calculations to
avoid confusion. More information is provided below.

Daniel Jacob wrote:

   For benchmark purposes I think that both UCX and Tropchem should
   use the simple SOA in hetchem, AOD, and PM2.5 calculations. I
   think that was the intent, it makes the benchmarks comparable,
   and it is in the spirit of the simple SOA being the default. The
   main reason for including the complex SOA in the benchmark, as I
   understand, is to monitor any changes in that chemistry for the
   benefit of users wanting complex SOA.

Colette Heald wrote:

   We recommended a switch to turn on/off Eloise's mechanism, but it
   appears that this chemistry is always on (I don't think we thought
   through at the time that such a switch would require a whole new
   set of chemical mechanism files, etc.) and the only difference is
   whether it is included in the AOD and PM2.5 calculations. We think
   that this might cause some confusion in the community (and the
   expert user can always go into the code and make these adjustments
   themselves). So we'd like to recommend removing this switch prior
   to the release of v11.02 (and have all the SOA species included
   in the PM2.5 and AOD calculations in the complex scheme).
   Obviously this is a somewhat superficial functional change, so it
   is less urgent.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>