Add ND49 timeseries diagnostic to correct ozone to 10m
Katie Travis created a diagnostic to correct daytime ozone values from the
lowest model layer, ~60m, to 10m.

   C(zC) = (1-Ra(z1,zC)vd(z1))C(z1)	            Eq. 1

where Ra(z1,zC) is the aerodynamic resistance between z1 and zC, and vd(z1)
is the ozone deposition velocity at z1, and C(z1) is the ozone concentration
at z1.

Ra(z1,zC) is calculated to the lowest model level in drydep_mod.F. We
recalculate Ra using z1 = 10 m, which is the height of the CASTNET
measurement for ozone. The new Ra is added to the diagnostic array AD_RA
and passed to diag49.F for use in Equation 1.

This new diagnostic is called O3@10m-$, and can be called with tracer 539
in ND49 in input.geos.

References

    Travis, K.R., D.J. Jacob, C.A. Keller, S. Kuang, J. Lin, M.J. Newchurch,
    A.M. Thompson, Resolving ozone vertical gradients in air quality models,
    Atmos. Chem. Phys. Disc.,2017.

    Zhang, L., D.J. Jacob, E.M. Knipping, N. Kumar, J.W. Munger, C.C.
    Carouge, A. van Donkelaar, Y. Wang, and D. Chen, Nitrogen deposition to
    the United States: distribution, sources, and processes, Atmos. Chem.
    Phys., 12, 4,539-4,4554, 2012.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>