Move ND55 (bpch) tropopause diagnostics from AIRQNT to main.F
In order to replicate the prior code, the ND55 diagnostics are now
updated in main.F (in the same location where the call to routine
DIAG_TROPOPAUSE was located).  This is because the ND55 diagnostics
were only updated at that point, but the TropLev and TropHt fields
of State_Met are updated in AIRQNT, which is called multiple times
per heartbeat timestep.  This move is to make sure we can replicate
the bpch diagnostic output from the prior commit.

In GeosCore/linoz_mod.F, now only compute L_OVERWLD if it is a
tagged O3 simulation.  It is not needed for a fullchem simulation.

Lastly, we have converted State_Met%TropLev from a REAL(fp) variable
to an INTEGER variable (similar to State_Met%ChemGridLev).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>