Fixed bug in error check for collection metadata in history_mod.F90
In routine GetCollectionMetaData, we have changed the DO loop:

  DO N = CollectionStart, CollectionCount

to

  DO N = 1, CollectionCount

because the former code was causing an error if the collections were
listed in a different order in HISTORY.rc than specified under the
COLLECTIONS section.  This might be a little less efficient, but because
this code only executes during initialization, it is probably not a
huge deal.  We might find a way to skip collections that we have
encountered previously at a later date.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>