Increase Input_Opt%MAX_SPC to 600 to avoid issues in TOMAS simulations
As of v11-02e, the TOMAS40 simulation has 500 advected species listed in
input.geos. To avoid issues in the future when adding species to the full-
chemistry simulations, we increase Input_Opt%MAX_SPC in main.F.

At some point, we should probably remove Input_Opt%MAX_SPC completely and
have GEOS-Chem read in and compute the total number of species at the very
beginning of the simulation.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>