Add a total collection to GEOS-Chem Classic History component
In addition to "instantaneous" and "time-averaged", the GEOS-Chem
"Classic" History component can now save diagnostic quantities
to "total" collections.

A "total" collection is similar to a "time-averaged" collection (where
we update into the accumulator array), except that we do not divide
by the number of updates.

The "total" collection is necessary in order to mimic the behavior
of certain GEOS-Chem "bpch" diagnostics, such as ND05.

NOTE: Appears to work but further validation may be necessary.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>