Add clarifying text to m_do_err_out.F90 netcdf error routine
We have added an additional sentence to the netCDF error output
to indicate that the error message was caused in either writing
to or reading from a netCDF file.

NOTE: We hope to have the ability to return the netCDF error to
the main program level in the near future.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>