Bypass initialization of diagnostics for commented out collection names
A collections linked list is now created when creating DiagList. It
includes collections that are uncommented in the COLLECTIONS section at
the top of HISTORY.rc. Diagnostics in collections that are not in the
collections linked list will be ignored completely.

Please note that this updates in the GCC history component are required
to make usage of commenting out collection names in the COLLECTIONS
section of HISTORY.rc. This feature is available in GCHP, however, and
can now be utilized with the new netcdf diagnostics.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>