Remove setting concentrations of O2, N2, and H2 from flexchem_mod.F90
In flexchem_mod.F90, we were setting concentrations of species O2, N2, and
H2 to specific values at each chemistry timestep. This was causing problems
for tropchem-based mechanisms following the removal of the UCX CPP switch
because ind_N2 was not defined. To resolve this, we simply need to define
N2 in species_database_mod.F90 and set BackgroundVV to the value previously
specified in flexchem_mod.F. This will ensure N2 has the correct concentrations
in State_Chm%Species when N2 is defined as a chemical species. Then at the
beginning of each chemistry timestep KPP's concentration array C is
overwritten by State_Chm%Species.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>