Minor bug fixes: set POA1 as hyg grth spc, shift NC_Diag ifdefs
If NC_DIAG=y and running the complex SOA simulation then POA1 is used in
place of OCPI for hygroscopic growth species diagnostics. This requires
the hygroscopic growth logical for POA1 to be set to true in the species
database. Not doing so results in a reduced set of hygroscopic growth
species since OCPI is not a species and therefore the run will fail.

An additional update in this commit is shifting the placement of the
NC_DIAG if defined statement to make all subroutines stubs if running
with binary diagnostics only.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>