Now compute aerosol mass and PM2.5 diagnostics in diagnostic_mod.F90
The ND42 bpch (aerosol mass and PM2.5) diagnostic is updated in
routine DIAG42, which is called at the end of a GEOS-Chem timestep.

We have implemented the corresponding subroutine (Set_AerMass_Diagnostic)
in module GeosCore/diagnostics_mod.F90 for the netCDF output.

May need further validation but at this point it compiles and runs.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>