Fixed parallelization error in the computation of State_Met%THETA
THETA was formerly computed diag1.F only as a diagnostic.  In v11-02
we introduced State_Met%THETA so that we could attach this to the
netCDF diagnostics.

State_Met%THETA needs to know the surface pressure.  We had used
State_Met%PEDGE(I,J,1).  But because of the parallel DO loop, we
cannot assume that State_Met%PEDGE(I,J,1) will be populated (if the
first level of the parallelization is different than 1).

The quick fix is to add a call to GET_PEDGE( I, J, 1 ) in the
computation of State_Met%THETA.  This will always return the
surface pressure no matter which level is done first in the parallel
loop.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>