Comment out rest of code ND49 timeseries diagnostic to correct ozone to 10m
Not commenting out the arrays for correcting ozone to 10m was causing issues
when users compiled with NC_DIAG=y and BPCH_DIAG=n. To avoid this, we comment
out the rest of the code for this diagnostic which is not fully supported yet.
Also, #ifdefs for BPCH_DIAG have been added where needed to avoid issues
when this diagnostic is reactivated.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>