Initialize LSTOT, LSTOT0 variables in routine CHEM_SO2
In the parallel loop in routine CHEM_SO2, the LSTOT and LSTOT0
variables can potentially be used before they are initialized.
Make sure to zero LSTOT and LSTOT0 at the top of the loop.  Also,
for safety's sake, zero LSTOT again in the ELSE block where several
other variables are zeroed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>