Change IBr/ICl/pFE names from NameAllCaps to IBr/ICl/pFe in spc database
Using these species name as all caps in the species database causes GCHP
to fail because it conflicts with the name in the advected species list of
input.geos and therefore the name used in the MAPL Internal State. Later
retrieval using the species database looks for 'SPC_IBR', 'SPC_ICL', and
'SPC_PFE' unsuccessfully.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>